Refactor: Rename BinomialConstraint.h with binomial_constraint.h
